requirements-vercel.txt
Flask==2.3.3
sentence-transformers==2.7.0
faiss-cpu==1.7.4
numpy==1.26.4
validators==0.20.0
gunicorn==20.1.0
torch==2.0.1+cpu
torchvision==0.15.2+cpu
torchaudio==2.0.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
pandas==2.2.2
tqdm==4.66.4 